let the linear solver backend determine the type of the sparse matrix adapter
so far, this property was specified by the level of the base
discretization. because explicitly set properties overwrite the stuff
inherited from splices, so far it was not possible to use a linear
solver backends which requires a custom sparse matrix adapter unless
the sparse matrix adapter adapter class was explicitly set again at
the problem level -- thus making the splice much less generic than it
ought to be. (also, the whole point of the sparse matrix adapter
abstraction is that the discretization does not need to know what it
does internally.)